Add missing SpaceVim files (Fix #14)
When `:DeinUpdate` encounters an error updating one of the plugins, it uses the SpaceVim logger to display the message. It seems not all of the required SpaceVim files were added to this repository for that to work correctly, unless the definitions were already loaded elsewhere (e.g., if the user is running the full SpaceVim installation). SpaceVim's `api/logger.vim` depends on `api/time.vim` because it includes the current time in the log message.